Refactor: each challenge code in separated files
